move .doc/ to utils/ to be able to have the helper script ./cv in root folder
