+++ .travis.yml
@@ -13,8 +13,9 @@ env:
   - CV_BUILD="test"
   - CV_BUILD="docs" ENCRYPTION_LABEL="4c79fec0aeb9" COMMIT_AUTHOR_EMAIL="generator@cometvisu.org" CV_BROWSER=Chrome CV_VERSION=stable
 before_install:
+- mkdir -p downloads
 - nvm install 4.2
-- if [ "$CV_BUILD" == "docs" ]; then sudo ./.doc/travis/browser-setup.sh; fi
+- if [ "$CV_BUILD" == "docs" ]; then sudo ./utils/travis/browser-setup.sh; fi
 - if [ "$CV_BUILD" == "docs" ]; then export CHROME_BIN=`pwd`/chrome/google-chrome; fi
 - if [ "$CV_BUILD" == "docs" ] && ! [ -e downloads/plantuml.deb ]; then wget -O downloads/plantuml.deb http://yar.fruct.org/attachments/download/362/plantuml_7707-1_all.deb; fi
 - mkdir cache
@@ -22,10 +23,10 @@ install:
 - npm install -g grunt-cli
 - npm install
 - "./node_modules/protractor/bin/webdriver-manager update"
-- if [ "$CV_BUILD" == "docs" ]; then sudo pip install -r .doc/requirements.txt; fi # for documentation
+- if [ "$CV_BUILD" == "docs" ]; then sudo pip install -r utils/requirements.txt; fi # for documentation
 - if [ "$CV_BUILD" == "docs" ]; then sudo apt-get install libxml2-dev libxslt-dev graphviz; fi
 - if [ "$CV_BUILD" == "docs" ]; then sudo dpkg -i downloads/plantuml.deb; fi
-- if [ "$CV_BUILD" == "docs" ]; then .doc/travis/npm-setup.sh; fi
+- if [ "$CV_BUILD" == "docs" ]; then utils/travis/npm-setup.sh; fi
 before_script:
 - export DISPLAY=:99.0
 - sh -e /etc/init.d/xvfb start
@@ -35,7 +36,7 @@ script:
 - if [ "$CV_BUILD" == "test" ]; then grunt jscs; fi
 - if [ "$CV_BUILD" == "test" ]; then grunt karma:travis; fi
 - if [ "$CV_BUILD" == "test" ]; then grunt e2e; fi
-- if [ "$CV_BUILD" == "docs" ]; then .doc/travis/deploy.sh; fi
+- if [ "$CV_BUILD" == "docs" ]; then utils/travis/deploy.sh; fi
 after_success:
 - if [ "$CV_BUILD" == "test" ]; then grunt coveralls; fi
