+++ Gruntfile.js
@@ -371,7 +371,7 @@ module.exports = function(grunt) {
         options: {
           destination: grunt.option('targetDir') || 'doc/api/html',
           template : "node_modules/ink-docstrap/template",
-          configure : ".doc/jsdoc.conf.json"
+          configure : "utils/jsdoc.conf.json"
         }
       },
       rst : {
@@ -383,7 +383,7 @@ module.exports = function(grunt) {
         options: {
           destination: grunt.option('targetDir') || 'doc/api/rst',
           template : "node_modules/jsdoc-sphinx/template/",
-          configure : ".doc/jsdoc.conf.json"
+          configure : "utils/jsdoc.conf.json"
         }
       }
     },
@@ -517,7 +517,7 @@ module.exports = function(grunt) {
       },
       screenshots: {
         options: {
-          configFile: ".doc/protractor.conf.js",
+          configFile: "utils/protractor.conf.js",
           args: {
             params: {
               subDir: grunt.option('subDir'),
@@ -532,7 +532,7 @@ module.exports = function(grunt) {
       },
       screenshotsSource: {
         options: {
-          configFile: ".doc/protractor.conf.js",
+          configFile: "utils/protractor.conf.js",
           args: {
             params: {
               subDir: "jsdoc"
@@ -546,7 +546,7 @@ module.exports = function(grunt) {
       },
       screenshotsManual: {
         options: {
-          configFile: ".doc/protractor.conf.js",
+          configFile: "utils/protractor.conf.js",
             args: {
             params: {
               subDir: "manual"
