+++ utils/plugins/widget_example.js
