+++ utils/travis/npm-setup.sh
